networking: Refactor request-response module.
