+++ crates/subspace-networking/src/protocols.rs
@@ -1,4 +1,4 @@
 pub(crate) mod connected_peers;
 pub mod peer_info;
-pub mod requests;
+pub mod request_response;
 pub(crate) mod reserved_peers;
