+++ crates/subspace-networking/src/protocols/request_response/handlers/generic_request_handler.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::protocols::requests::request_responses::{
+use crate::protocols::request_response::request_response_factory::{
     IncomingRequest, OutgoingResponse, ProtocolConfig, RequestHandler,
 };
 use async_trait::async_trait;
