+++ crates/subspace-networking/src/protocols/request_response/handlers/piece_by_index.rs
