+++ crates/subspace-networking/src/protocols/request_response/request_response_factory.rs
@@ -187,7 +187,7 @@ pub struct OutgoingResponse {
     pub sent_feedback: Option<oneshot::Sender<()>>,
 }
 
-/// Event generated by the [`RequestResponsesBehaviour`].
+/// Event generated by the [`RequestResponseFactory`].
 #[derive(Debug)]
 pub enum Event {
     /// A remote sent a request and either we have successfully answered it or an error happened.
@@ -205,7 +205,7 @@ pub enum Event {
         result: Result<(), ResponseFailure>,
     },
 
-    /// A request initiated using [`RequestResponsesBehaviour::send_request`] has succeeded or
+    /// A request initiated using [`RequestResponseFactory::send_request`] has succeeded or
     /// failed.
     ///
     /// This event is generated for statistics purposes.
@@ -264,9 +264,10 @@ impl IfDisconnected {
     }
 }
 
-/// Implementation of `NetworkBehaviour` that provides support for request-response protocols.
+/// Implementation of `NetworkBehaviour` that provides support for multiple request-response
+/// protocols.
 #[allow(clippy::type_complexity)] // to preserve compatibility with copied implementation
-pub struct RequestResponsesBehaviour {
+pub struct RequestResponseFactory {
     /// The multiple sub-protocols, by name.
     /// Contains the underlying libp2p `RequestResponse` behaviour, plus an optional
     /// "response builder" used to build responses for incoming requests.
@@ -314,7 +315,7 @@ struct RequestProcessingOutcome {
     response: OutgoingResponse,
 }
 
-impl RequestResponsesBehaviour {
+impl RequestResponseFactory {
     /// Creates a new behaviour. Must be passed a list of supported protocols. Returns an error if
     /// the same protocol is passed twice.
     pub fn new(
@@ -413,7 +414,7 @@ impl RequestResponsesBehaviour {
     }
 }
 
-impl NetworkBehaviour for RequestResponsesBehaviour {
+impl NetworkBehaviour for RequestResponseFactory {
     type ConnectionHandler = MultiHandler<
         String,
         <RequestResponse<GenericCodec> as NetworkBehaviour>::ConnectionHandler,
