--- crates/subspace-networking/src/protocols/requests/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod handlers;
-pub(crate) mod request_responses;
